all the rest of chapter 4

+++ osa4/blogs/controllers/blogs.js
@@ -1,31 +1,49 @@
+const jwt = require("jsonwebtoken");
 const blogsRouter = require("express").Router();
-const e = require("express");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
-  if (!body.title || !body.url) {
-    response.status(400).end();
-  } else {
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      likes: body.likes || 0,
-      url: body.url,
-    });
-    const blogSaved = await blog.save();
-    response.status(201).json(blogSaved);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
   }
+  const user = await User.findById(decodedToken.id);
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    likes: body.likes || 0,
+    url: body.url,
+    user: user._id,
+  });
+  const blogSaved = await blog.save();
+  user.blogs = user.blogs.concat(blogSaved._id);
+  await user.save();
+  response.status(201).json(blogSaved);
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+  if (user._id.toString() === blog.user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({
+      error:
+        "not same person trying to delete the blog as the person who created it",
+    });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response, next) => {
